Clase 12 politica de acceso proteger eliminar

+++ app/Http/Controllers/RepositoryController.php
@@ -37,7 +37,11 @@ class RepositoryController extends Controller
 
     }
 
-    public function destroy(Repository $repository){
+    public function destroy(Request $request, Repository $repository){
+
+        if($request->user()->id !== $repository->user_id){
+            abort(403);
+        }
 
         $repository()->delete();
